Move Menu to future internally
