+++ .changeset/loud-terms-help.md
@@ -1,0 +1,5 @@
+---
+'@kaizen/components': patch
+---
+
+Move Menu code to future internally
