+++ packages/components/src/Menu/_docs/Menu.stories.tsx
@@ -8,7 +8,7 @@ import { MenuItem } from '../subcomponents/MenuItem'
 import { MenuList } from '../subcomponents/MenuList'
 
 const meta = {
-  title: 'Actions/Menu/v1',
+  title: 'Components/Menu',
   component: Menu,
   args: {
     button: (
