+++ packages/components/src/__actions__/Menu/v3/index.ts
@@ -1,6 +1,2 @@
-export * from './Menu'
-export * from './MenuItem'
-export * from './MenuTrigger'
-export * from './MenuPopover'
-export * from './MenuSection'
-export * from './MenuHeader'
+// just here temporarily until we remove all v3 usages
+export * from '../../../__future__/Menu'
