+++ packages/components/src/__future__/Menu/_docs/ApiSpecification.mdx
@@ -3,28 +3,27 @@ import { KAIOInstallation, ResourceLinks } from '~storybook/components'
 import * as docsStories from './Menu.docs.stories'
 import * as exampleStories from './Menu.stories'
 
-<Meta title="Actions/Menu/v3/API Specification" />
+<Meta title="Components/Menu/Menu (Future)/API Specification" />
 
-# Menu API Specification (v3)
+# Menu API Specification (Future)
 
 Updated December 16, 2024
 
 <ResourceLinks
   sourceCode="https://github.com/cultureamp/kaizen-design-system/tree/main/packages/components/src/__actions__/Menu"
   figma="https://www.figma.com/design/eZKEE5kXbEMY3lx84oz8iN/%F0%9F%92%9C-Heart-UI-Kit?m=auto&node-id=6262-0&t=I4H23jCrCJz4O2nf-1"
-  designGuidelines="/?path=/docs/actions-menu-v3-usage-guidelines--docs"
+  designGuidelines="/?path=/docs/"
   ariaComponentPage="https://react-spectrum.adobe.com/react-aria/Menu.html"
 />
 
 <KAIOInstallation
   exportNames={['Menu', 'MenuTrigger', 'MenuItem', 'MenuPopover', 'MenuSection', 'MenuHeader']}
-  family="actions"
-  version="3"
+  isFuture
 />
 
 ## Compatibility
 
-The `Menu` `v3` component is not backwards compatible with Kaizen `Button` `v1` and `v2`. This will only work when used with the `Button` `v3` component.
+The `Menu (Future)` component is not backwards compatible with Kaizen `Button` `v1` and `v2`. This will only work when used with the `Button` `v3` component.
 
 ## Overview
