+++ packages/components/src/__future__/Menu/_docs/Menu.mdx
@@ -3,9 +3,9 @@ import { ResourceLinks, Installation, DosAndDonts, DoOrDont } from '~storybook/c
 import * as MenuDocsStories from './Menu.docs.stories'
 import * as MenuStories from './Menu.stories'
 
-<Meta title="Actions/Menu/v3/Usage Guidelines" />
+<Meta title="Components/Menu/Menu (Future)/Usage Guidelines" />
 
-# Menu (v3)
+# Menu (Future)
 
 Updated July 6, 2024
 
@@ -18,7 +18,7 @@ Updated July 6, 2024
 
 <Installation
   installCommand="pnpm add @kaizen/components"
-  importStatement='import { MenuTrigger, Menu, MenuItem, MenuPopover, MenuSection, MenuHeader } from "@kaizen/components/v3/actions"'
+  importStatement='import { MenuTrigger, Menu, MenuItem, MenuPopover, MenuSection, MenuHeader } from "@kaizen/components/future"'
 />
 
 ## Overview
